processing/virustotal/requirements.txt
virustotal-api==1.1.10
